refactor: extract publishDisplayInfo
 - prune unused type imports
 - FastLP token name is decided; prune UNTIL
